module-2/session-9/next-day-calculator/src/NextDayCalculator.java
public class NextDayCalculator {
    public static String getNextDay(int day, int month, int year) {
        int lastDayOfMonth = getLastDayOfMonth(month, year);

        if (day == lastDayOfMonth) {
            day = 1;
            if (month == 12) {
                month = 1;
                year++;
            } else {
                month++;
            }
        } else {
            day++;
        }

        return day + "/" + month + "/" + year;
    }

    private static int getLastDayOfMonth(int month, int year) {
        switch (month) {
            case 4:
            case 6:
            case 9:
            case 11:
                return 30;
            case 2:
                return isLeapYear(year) ? 29 : 28;
            default:
                return 31;
        }
    }

    private static boolean isLeapYear(int year) {
        if (year % 400 == 0) return true;
        if (year % 100 == 0) return false;
        return year % 4 == 0;
    }
}
